+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `inline-confirm` will be documented in this file.
 
+## v0.2.3 - 2026-06-04
+
+- chore: refactor IDE stub generator script
+
+## v0.2.2 - 2026-06-04
+
+- fix: update stub file to follow Filament return type pattern
+
+## v0.2.1 - 2026-06-04
+
+- feat: add IDE helper stubs
+
 ## v0.2.0 - 2026-06-03
 
 - feat: support Action inside an ActionGroup (dropdown menus and button groups)